// src/services/types.ts
export interface CurrencyRate {
  code: string
  rate: number
  flag: string
  change?: number
  name?: string
}

export interface CurrencyConversion {
  fromCurrency: string
  toCurrency: string
  amount: number
  rate: number
  result: number
  timestamp: string
}

export interface ExchangeRateResponse {
  base: string
  date: string
  rates: { [key: string]: number }
}